requirements.txt
fastapi
uvicorn
websockets
python-dotenv
aiohttp
assemblyai>=0.20.0  # Swapped from deepgram-sdk
requests
pydantic
textblob
pyaudio
